Add dvm upgrade command
This allows dvm to self-update to the latest version. We do our best
to avoid problems when local patches exist by requesting the user commit
or stash them.

It seems reasonable that a user may have local commits that for whatever
reason they don't want to upstream, or whose patch was rejected. In
those cases, we do our best to merge changes. In the case of local
commits that conflict with updates from upstream, we notify the user,
then abort the git pull --rebase to try to get things back into a
working state.