Use --exit-code flag in diff
Due to a reversed conditional, _dvm_needsupgrade always returned 1 even
when no upgrade was needed. This fixes the conditional and runs git diff
with the --exit-code option to ensure we get the same exit code
behaviour as regular diff; in other words, exit with 1 if there are
diffs and 0 otherwise.