Support downloading Dartium on arm64 Macs
The chance that anyone is actually using Dartium or Content Shell, and
thus Dart 1.x in 2022 is pretty slim, but adding this for completionism
and for the fun factor of jumping in the time machine and paying a visit
to Dart's origins in Chromium.

Older versions of Dartium were time-limited to stop working after some
fixed amount of time due to the security implications of running an
out-of-date browser with two VMs running simultaneously and memory leaks
galore. My recollection is that later versions of Dartium removed this
time limit in order to give Dart 1.x users time to migrate to Dart 2, so
those are likely usable even today.