Use arm64 SDK on arm64 Macs where available (#39)
Dart SDK 2.14.1 and later are available as arm64 binaries for MacOS, for use on
Apple Silicon Macs.

With this change, dvm will download the arm64 Mac version of the Dart SDK for
all versions where it is available when running on an arm64 Mac. Where an arm64
build is not available we continue to download the x64 or (for older versions
of Dartium and Content Shell) an ia32 build.